Added new page to demo app. Switch error message to use message and not i18nkey
