+++ addon/components/item-picker/item-preview/template.hbs
@@ -1,6 +1,6 @@
 {{#if showError}}
   <p class="alert {{messageClass}}">
-    {{errorMessage.i18nKey}}
+    {{errorMessage.message}}
   </p>
 {{/if}}
 {{#unless showFallback}}
