+++ tests/dummy/app/application/template.hbs
@@ -18,6 +18,7 @@
               {{#active-link}}{{link-to "Default Usage" "itempicker.index"}}{{/active-link}}
               {{#active-link}}{{link-to "Multi-Select" "itempicker.multiselect"}}{{/active-link}}
               {{#active-link}}{{link-to "Facets" "itempicker.facets"}}{{/active-link}}
+              {{#active-link}}{{link-to "Validation" "itempicker.validation"}}{{/active-link}}
           </ul>
         </li>
       </ul>
